разбил код на модули по задачам
