+++ textures_tool.pyproj
@@ -23,6 +23,11 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="main.py" />
+    <Compile Include="img_manipulation.py" />
+    <Compile Include="csv_manipulation.py" />
+    <Compile Include="assist_funcs.py" />
+    <Compile Include="window.py" />
+    <Compile Include="buffer.py" />
   </ItemGroup>
   <ItemGroup>
     <Interpreter Include="..\imgs_numbering\env\">
